feat. 당첨 번호를 입력한다.
