+++ README.md
@@ -47,7 +47,7 @@
 ## Todo
 - [x] 구매 금액을 입력한다. 가격에 따른 구매 갯수를 저장하고, 출력한다
 - [x] 14개의 랜덤 결과를 출력한다.
-- [ ] 당첨 번호를 입력한다.
+- [x] 당첨 번호를 입력한다.
 - [ ] 당첨 통계를 출력한다.
 - [ ] 수익률을 출력한다.
