+++ src/main/java/step2/InputView.java
@@ -1,8 +1,10 @@
 package step2;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
+import java.util.stream.Collectors;
 
 public class InputView {
     public static final String BUY_TEXT = "구입금액을 입력해 주세요.";
@@ -13,13 +15,9 @@ public class InputView {
         return sc.nextInt();
     }
 
-    public static void printLottoCount(int lottoNum) {
-        System.out.println(lottoNum + "개를 구매했습니다.");
-    }
-
-    public static void printLottoNum(ArrayList<List<Integer>> lottoArray) {
-        for (List<Integer> lotto : lottoArray) {
-            System.out.println(lotto.toString());
-        }
+    public static List<Integer> getLastWinner() {
+        sc.nextLine();
+        System.out.println("\n지난 주 당첨 번호를 입력해 주세요");
+        return Arrays.stream(sc.nextLine().replace(" ", "").split(",")).mapToInt(Integer::parseInt).boxed().collect(Collectors.toList());
     }
 }
